fix(packaging): ship the demo as 'netcanon demo' (wheel + image)
Finding R-02 / CF-02 / DA-01 from the 2026-06-06 review. The README's 'See it in 10 seconds' command ran tools/demo.py, but repo-root tools/ is excluded from both the wheel (packages.find = netcanon* only) and the Docker image (Dockerfile copies only netcanon/), and the image entrypoint is the server — so 'docker run ... python tools/demo.py' could never work against a published artifact.

Moved the demo scenarios into the package (netcanon/tools/demo.py) behind a new 'netcanon demo' subcommand that ships automatically (verified: the built wheel now contains netcanon/tools/demo.py + cli.py). Repo-root tools/demo.py kept as a thin re-export shim so 'python tools/demo.py' still works from a source checkout (zero churn to the ~15 walkthrough/AGENTS references). README hero corrected to 'docker run --rm --entrypoint netcanon ... demo --pair cisco__junos'; AGENTS doc-sync rows + tools/README repointed at the new scenario location.

Tests: tests/unit/tools/test_demo.py (+9: module, CLI delegation, shim re-export). Full unit suite green.
